+++ __init__.py
@@ -2,10 +2,10 @@ bl_info = \
     {
         "name": "Blender NLA Midi Copier",
         "author": "Cornerback24",
-        "version": (0, 18, 0),
+        "version": (0, 19, 0),
         "blender": (2, 80, 0),
-        "location": "View NLA Editor > Tool Shelf",
-        "description": "Copy actions to action strips based on midi file input",
+        "location": "NLA Editor and Graph Editor Sidebar",
+        "description": "Copy actions to action strips and generate graph editor keyframes based on midi file input",
         "doc_url": "https://github.com/Cornerback24/Blender-NLA-Midi-Copier#blender-nla-midi-copier",
         "tracker_url": "https://github.com/Cornerback24/Blender-NLA-Midi-Copier/issues",
         "support": "COMMUNITY",